LayoutTests/http/wpt/user-activation/bless.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>User activation: bless works in iframe</title>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
    <iframe id="same-origin"></iframe>
    <iframe id="cross-origin"></iframe>
</body>
<script type="module">
    const iframeSameOrigin = document.querySelector("iframe#same-origin");
    const iframeCrossOrigin = document.querySelector("iframe#cross-origin");

    function sendMessage(iframe) {
        return new Promise((resolve, reject) => {
            window.addEventListener("message", function messageListener(event) {
                if (event.source === iframe.contentWindow) {
                    window.removeEventListener("message", messageListener);
                    resolve(event.data);
                }
            });
            iframe.contentWindow.postMessage("run", "*");
        });
    }

    function loadIframe(iframe, src) {
        return new Promise((resolve) => {
            iframe.onload = () => resolve();
            iframe.src = src;
        });
    }

    promise_setup(async () => {
        const hostInfo = get_host_info();
        const sameOriginSrc = document.location.pathname.replace(
            /bless.html$/,
            "resources/iframe.html"
        );
        const crossOriginSrc = new URL(
            sameOriginSrc,
            hostInfo.HTTPS_REMOTE_ORIGIN
        ).href;

        await Promise.all([
            loadIframe(iframeCrossOrigin, crossOriginSrc),
            loadIframe(iframeSameOrigin, sameOriginSrc),
        ]);
    });

    promise_test(async (t) => {
        await test_driver.bless("user activation");
    }, "bless in top document");

    promise_test(async (t) => {
        await test_driver.bless(
            "user activation blessed directly from parent script",
            null,
            iframeSameOrigin.contentWindow
        );
    }, "bless added directly from parent script to same-origin iframe's window context");

    promise_test(async (t) => {
        const result = await sendMessage(iframeSameOrigin);
        assert_equals(
            result,
            "success",
            "bless in same-origin iframe should succeed"
        );
    }, "bless added from parent script to same-origin iframe via sendMessage");

    promise_test(async (t) => {
        const result = await sendMessage(iframeCrossOrigin);
        assert_equals(
            result,
            "success",
            "bless in cross-origin iframe should succeed"
        );
    }, "bless added from inside cross-origin iframe");
</script>
